+++ Changes.md
@@ -1,3 +1,37 @@
+Change Log in version 403.1.6 (2023111807)
+===========================================
+1. Fix missing 'recaptcha' on login form.
+2. Only show secondary navigation if there is something to show.
+3. Refactored 'buttonfocuscolor' and 'buttonfocuscoloropacity' to be 'inputbuttonfocuscolour' and 'inputbuttonfocuscolouropacity'
+   respectively as they really relate to the box shadow when focusing on an 'input' button.
+4. Created 'buttonfocuscolour' and 'buttontextfocuscolor' to apply a background and text colour on focused buttons.
+5. Flex Slider upgraded to 2.7.2.
+6. Pace updated to 1.2.4.
+7. AMD'ify Tickerme.
+8. AMD'ify jquery.easing.
+9. Removed outdated method of jQuery plugin usage.
+10. Moved ticker JS to sponsors only.
+11. Added 'enablesavecanceloverlayresetconfirm' string for save settings cancellation reset warning.
+12. Fix sticky header margin.
+13. Added 'regionmaintextcolor' and other small fixes.
+14. Added 'primarycolour' and 'secondarycolour', to control the respective Bootstrap colours as used in the SCSS in general.
+15. Added the ability to set text and background colours on the Ticker.
+16. Moved 'fontcolor' to 'Colours' tab and fix button colours.
+17. Dialog and tab tidy.
+18. Ticker style back to what it was.
+19. Fix collapsed information blocks not showing.
+20. Remove 'docked' styles.
+21. Fix 'Flexible blocks are too big on small screens'.
+22. Add name and description back to user profile page.
+23. Fix theme selector buttons.
+24. Tidy frontpage layout.
+25. Fix backup 'Return to course button'.
+26. Add file settings to import / export of settings.
+27. Fix unread notification text colour.
+28. Fix notification icons not showing when body font weight is <= 300.
+29. Update theme version of FontAwesome free to 6.6.0 from 6.4.2.
+30. Increase tooltip font size to .9rem.
+
 Change Log in version 403.1.5 (2023111806)
 ===========================================
 1. Reduce drawer flicker on page load when the navbar is sticky.